Add a filter for generated HTML
JavaScript is not run during critical CSS generation, which may
result in some unexpected HTML (e.g. a missing class name added by
a script to manage font loading)

This allows us to filter HTML on an invidual request to modify
content as it may be seen in a browser with JavaScript.